duplicate entries, 'sound' multimode fix, pretty CLI.
Originally AudioSystem stops creating duplicate entries, which led to, same filename different file extention, files to not be loaded as it's 'technically' already loaded based on the filename (not filename.extension).
AudioSystem now allows duplicate file names to be loaded but with a unique number prefixed to the duplicated entry with the attatched filepath to the file with different file extension.
This change is also reflected on the PySoundboard GenerateSoundIndex to properly reference the duplicated entry.

SoundEffect fix for multimode/singlemode, it broke when i was implementing the feature to AudioMedia.

The CLI output is now pretty-fied. Which I will probably regret later but im pushing it anyways :))))))))))