Pulling Latest Stable (#3)
* Added Workaround for my Linux Pipewire problems.

I was having problems with frontpanel mic --> headset while having my soundboard --> both my mic mix and headset in Pipewire.
The Soundboard inadvertantly "connected" the mic and front-panel, which causes some weird stuttering and crunchy audio from front-panel while I have Firefox playing audio.

so I added a linux-exclusive patch.
Added hostAudioPool, which functions the same as audioPool['audio'], but uses the default audio device.
This way you can still hear the sound while also still playing through the microphone without manual Pipewire re-routing.

* duplicate entries, 'sound' multimode fix, pretty CLI.

Originally AudioSystem stops creating duplicate entries, which led to, same filename different file extention, files to not be loaded as it's 'technically' already loaded based on the filename (not filename.extension).
AudioSystem now allows duplicate file names to be loaded but with a unique number prefixed to the duplicated entry with the attatched filepath to the file with different file extension.
This change is also reflected on the PySoundboard GenerateSoundIndex to properly reference the duplicated entry.

SoundEffect fix for multimode/singlemode, it broke when i was implementing the feature to AudioMedia.

The CLI output is now pretty-fied. Which I will probably regret later but im pushing it anyways :))))))))))